+++ lib/entities/settlements.dart
@@ -1,3 +1,20 @@
+
 export '../src/entities/settlements/goods.dart' show Goods;
 export '../src/entities/settlements/location.dart' show Location;
 export '../src/entities/settlements/settlement.dart' show Settlement;
+
+export '../src/generators/settlements/locations/location_generator.dart'
+    show LocationGenerator;
+
+export '../src/generators/settlements/settlement_generator.dart'
+    show SettlementGenerator;
+
+export '../src/subtypes/settlements/settlement_type.dart' show SettlementType;
+export '../src/subtypes/settlements/settlements_manager.dart'
+    show SettlementManager;
+
+export '../src/subtypes/settlements/hamlet/hamlet.dart' show Hamlet;
+export '../src/subtypes/settlements/village/village.dart' show Village;
+export '../src/subtypes/settlements/town/town.dart' show Town;
+export '../src/subtypes/settlements/city/city.dart' show City;
+export '../src/subtypes/settlements/metropolis/metropolis.dart' show Metropolis;