+++ lib/src/subtypes/races/dragonborn/dragonborn.dart
@@ -1,8 +1,11 @@
-import '../../../../generators/generators.dart';
-import '../../../../subtypes/races.dart';
+import '../../../../entities/races.dart';
 import '../../../entities/npcs/alignment.dart';
 import '../../../entities/npcs/hair.dart';
 import '../../../enums/gender.dart';
+import '../../../generators/base/generator.dart';
+import '../../../generators/base/list_item_generator.dart';
+import '../../../generators/base/multiple_generator.dart';
+import '../../../generators/base/number_generator.dart';
 import '../../../generators/npcs/personality/alignment/base_alignment_generator.dart';
 import '../../../generators/npcs/personality/quirks/base_personality_quirk_generator.dart';
 import '../../../generators/npcs/personality/traits/personality_trait_generator.dart';